draft redis function to add users
