+++ lib/index.js
@@ -1,24 +1,14 @@
-require('env2')('config.env')
+require('env2')('./config.env')
+const createServer = require('./server.js')
+const createClient = require('./redis/client.js')
 
-const Hapi = require('hapi')
-const server = new Hapi.Server()
-
-server.connection({ port: process.env.PORT || 4000 })
-
-const routes = [{
-    method: 'GET',
-    path: '/',
-    handler: (request, reply) => {
-      reply('hello server')
-    }
-  }
-]
-
-server.route(routes)
+const client = createClient()
+const server = createServer(client)
 
 server.start((err) => {
-  if (err) throw err
-  else {
-    console.log('server is listening on port: ' + server.info.port)
+  if (err) {
+    console.log('server error: ', err)
+  } else {
+    console.log('server listening on port: ' + server.info.port)
   }
 })
