+++ lib/redis/redisFunctions.js
@@ -1,0 +1,6 @@
+// All database helper functions should take the redis client object as the first argument
+
+module.exports.addUser = (client, messageObject, credit) => {
+  const phoneNumber = messageObject.originator
+  return client.HMSET('users', phoneNumber, credit)
+}
